requirements.txt
requests
sanic
responses
pytest-cov
